Extract view doctor details dialog
